fix(eat): forks not correctly distributed
Forks were not being distributed every 2 neighbours. which could lead to
sub-optimal placement and otherwise avoidable philo death.

A tipical example of this is considering philos A, B, C and D:
- A and D start eating, blocking B and C
- A and D finish eating
- B start eating, blocking C, A and D are sleeping

A more optimal placement is:
- A and C starts eating, blocking B and C
- A and C finish eating
- B and D start eating, A and C are sleeping

Notice how, in the second case, both remaining philos are able to eat,
contrarily to the first case.