perf(table_request): no need to update neighbours
Removed the need to recalculate neighbours eich time the table_request
fucntion is called: philo neighbours are now saved at table creation as
part of the s_philo struct. This means that philos are now aware of
their neighbours. Please be aware that this SHOULD NOT be used as a
means to communicate between philos. All philo communication is to
remain handled by the monitor (table).