fix(death): philos dying due to ordering
Reinstored old last_meal bias system in philo_init function so as to
favor even numbered philos. This is done by offsetting last_meal
for odd philos by the time it takes a philo to eat. Therefore, odd
philos are considered as having already eaten at the start of the
simulation. The result is that philos will alternate meals, which
avoids blockages.