perf(time): time is now sampled from table loop
Time is no longuer sampled on a per-thread basis but is instead
calculated by the monitor (table) as it loops over each thread to
determine their state. This avoids having to rely on multiple system
calls, which should speed up ressource requests significantly, but has
the drawback that the sampled time must be thread-safe (mutexes) and might be
less accurate.