fix(flush): race conditions when flushing messages
Fixed race conditions due to simultaneous read/write access of the
message bus when flushing. The resulting code is more blocking and
guarantees more security, but might be slower. Still, an effort has been
made to minimise the blocking on write when reading the buffer, so the
impact should be negligeable.